Move GrpcMessageInterceptor into auth package to prevent circular dependency.
