+++ core/src/main/java/feast/core/config/CoreSecurityConfig.java
@@ -29,7 +29,8 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @Slf4j
-@ComponentScan(basePackages = {"feast.auth.config", "feast.auth.service"})
+@ComponentScan(
+    basePackages = {"feast.auth.config", "feast.auth.service", "feast.auth.interceptors"})
 public class CoreSecurityConfig {
 
   /**
